remove pch / include what i need

+++ src/core/app.hpp
@@ -7,55 +7,55 @@
 #include "engine_window.hpp"
 
 namespace vke {
-    class App {
-    public:
-        App();
+class App {
+   public:
+	App();
 
-        ~App();
+	~App();
 
-        App(const App&) = delete;
+	App(const App&) = delete;
 
-        App& operator=(const App&) = delete;
+	App& operator=(const App&) = delete;
 
-        static constexpr int HEIGHT = 600;
-        static constexpr int WIDTH = 800;
+	static constexpr int HEIGHT = 600;
+	static constexpr int WIDTH = 800;
 
-        void run();
+	void run();
 
-    private:
-        void loadGameObjects();
+   private:
+	void loadGameObjects();
 
-        void createPipelineLayout();
+	void createPipelineLayout();
 
-        void createPipeline();
+	void createPipeline();
 
-        void createCommandBuffers();
+	void createCommandBuffers();
 
-        void drawFrame();
+	void drawFrame();
 
-        void recreateSwapChain();
+	void recreateSwapChain();
 
-        void recordCommandsBuffers(size_t imageIndex) const;
+	void recordCommandsBuffers(size_t imageIndex) const;
 
-        void renderGameObjects(const VkCommandBuffer* commandBuffer) const;
+	void renderGameObjects(const VkCommandBuffer* commandBuffer) const;
 
-        void freeCommandBuffers() const;
+	void freeCommandBuffers() const;
 
-        VkEngineWindow mVkWindow{WIDTH, HEIGHT, "VkEngine"}; // Vulkan window
+	VkEngineWindow mVkWindow{WIDTH, HEIGHT, "VkEngine"};  // Vulkan window
 
-        VkEngineDevice mVkDevice{mVkWindow};
+	VkEngineDevice mVkDevice{mVkWindow};
 
-        std::shared_ptr<VkEngineSwapChain> mVkSwapChain = nullptr;
-        std::unique_ptr<VkEnginePipeline> pVkPipeline = nullptr;
-        std::vector<VkEngineGameObjects> mVkGameObjects{};
+	std::shared_ptr<VkEngineSwapChain> mVkSwapChain = nullptr;
+	std::unique_ptr<VkEnginePipeline> pVkPipeline = nullptr;
+	std::vector<VkEngineGameObjects> mVkGameObjects{};
 
-        VkPipelineLayout pVkPipelineLayout = VK_NULL_HANDLE;
+	VkPipelineLayout pVkPipelineLayout = VK_NULL_HANDLE;
 
-        struct CommandBuffer {
-            VkCommandBuffer* ppVkCommandBuffers{};
-            u8 mSize{}; // number of command buffers (could be uint8_t)
-        } mCommandBuffer{};
+	struct CommandBuffer {
+		VkCommandBuffer* ppVkCommandBuffers{};
+		u32 mSize{};  // number of command buffers (could be uint8_t)
+	} mCommandBuffer{};
 
-        static inline u64 mCurrentFrame = 0;
-    };
-} // namespace vke
+	static inline u64 mCurrentFrame = 0;
+};
+}  // namespace vke
